# Conflicts:
#	cmake/features.cmake
#	src/bindings/python/src/pyopenvino/CMakeLists.txt
#	src/bindings/python/src/pyopenvino/graph/types/element_type.cpp
#	src/bindings/python/tests/mock/pyngraph_fe_mock_api/CMakeLists.txt
#	src/frontends/common/src/utils.hpp
#	tools/mo/openvino/tools/mo/main.py

+++ src/common/transformations/src/transformations/op_conversions/convert_mvn1_to_mvn6.cpp
@@ -27,7 +27,7 @@ ngraph::pass::ConvertMVN1ToMVN6::ConvertMVN1ToMVN6() {
         if (!input_rank.is_static()) {
             return false;
         }
-        int64_t start_axis = 1 + (!mvn_node->get_across_channels());
+        int64_t start_axis = 1 + static_cast<int64_t>(!mvn_node->get_across_channels());
         if (input_rank.get_length() <= start_axis) {
             return false;
         }
@@ -37,7 +37,7 @@ ngraph::pass::ConvertMVN1ToMVN6::ConvertMVN1ToMVN6() {
         auto mvn6_node = std::make_shared<ngraph::opset6::MVN>(input,
                                                                axes,
                                                                mvn_node->get_normalize_variance(),
-                                                               mvn_node->get_eps(),
+                                                               static_cast<float>(mvn_node->get_eps()),
                                                                ngraph::op::MVNEpsMode::OUTSIDE_SQRT);
 
         mvn6_node->set_friendly_name(mvn_node->get_friendly_name());
